+++ daml-assistant/src/DAML/Assistant/Tests.hs
@@ -15,6 +15,7 @@ import DAML.Project.Consts hiding (getDamlPath, getProjectPath)
 import System.Directory
 import System.Environment
 import System.FilePath
+import System.Info.Extra (isWindows)
 import System.IO.Temp
 import System.IO.Extra
 import Data.List.Extra
@@ -31,7 +32,6 @@ import Control.Monad
 import Conduit
 import qualified Data.Conduit.Zlib as Zlib
 import qualified Data.Conduit.Tar as Tar
-import qualified Data.SemVer as V
 
 -- unix specific
 import System.PosixCompat.Files (createSymbolicLink)
@@ -149,11 +149,11 @@ testGetSdk = Tasty.testGroup "DAML.Assistant.Env.getSdk"
                 expected2 = base </> "sdk"
 
             createDirectory expected2
-            (Just (SdkVersion got1), Just (SdkPath got2)) <-
+            (Just got1, Just (SdkPath got2)) <-
                 withEnv [ (sdkVersionEnvVar, Just expected1)
                         , (sdkPathEnvVar, Just expected2)
                         ] (getSdk damlPath projectPath)
-            Tasty.assertEqual "sdk version" expected1 (V.toString got1)
+            Tasty.assertEqual "sdk version" expected1 (versionToString got1)
             Tasty.assertEqual "sdk path" expected2 got2
 
     , Tasty.testCase "getSdk determines DAML_SDK from DAML_SDK_VERSION" $ do
@@ -165,11 +165,11 @@ testGetSdk = Tasty.testGroup "DAML.Assistant.Env.getSdk"
 
             createDirectoryIfMissing True (base </> "daml" </> "sdk")
             createDirectory expected2
-            (Just (SdkVersion got1), Just (SdkPath got2)) <-
+            (Just got1, Just (SdkPath got2)) <-
                 withEnv [ (sdkVersionEnvVar, Just expected1)
                         , (sdkPathEnvVar, Nothing)
                         ] (getSdk damlPath projectPath)
-            Tasty.assertEqual "sdk version" expected1 (V.toString got1)
+            Tasty.assertEqual "sdk version" expected1 (versionToString got1)
             Tasty.assertEqual "sdk path" expected2 got2
 
     , Tasty.testCase "getSdk determines DAML_SDK_VERSION from DAML_SDK" $ do
@@ -181,11 +181,11 @@ testGetSdk = Tasty.testGroup "DAML.Assistant.Env.getSdk"
 
             createDirectory expected2
             writeFileUTF8 (expected2 </> sdkConfigName) ("version: " <> expected1 <> "\n")
-            (Just (SdkVersion got1), Just (SdkPath got2)) <-
+            (Just got1, Just (SdkPath got2)) <-
                 withEnv [ (sdkVersionEnvVar, Nothing)
                         , (sdkPathEnvVar, Just expected2)
                         ] (getSdk damlPath projectPath)
-            Tasty.assertEqual "sdk version" expected1 (V.toString got1)
+            Tasty.assertEqual "sdk version" expected1 (versionToString got1)
             Tasty.assertEqual "sdk path" expected2 got2
 
     , Tasty.testCase "getSdk determines DAML_SDK and DAML_SDK_VERSION from project config" $ do
@@ -201,11 +201,11 @@ testGetSdk = Tasty.testGroup "DAML.Assistant.Env.getSdk"
             writeFileUTF8 (base </> "project" </> projectConfigName)
                 ("project:\n  sdk-version: " <> expected1)
             createDirectory expected2
-            (Just (SdkVersion got1), Just (SdkPath got2)) <-
+            (Just got1, Just (SdkPath got2)) <-
                 withEnv [ (sdkVersionEnvVar, Nothing)
                         , (sdkPathEnvVar, Nothing)
                         ] (getSdk damlPath projectPath)
-            Tasty.assertEqual "sdk version" expected1 (V.toString got1)
+            Tasty.assertEqual "sdk version" expected1 (versionToString got1)
             Tasty.assertEqual "sdk path" expected2 got2
 
 
@@ -223,11 +223,11 @@ testGetSdk = Tasty.testGroup "DAML.Assistant.Env.getSdk"
                 ("project:\n  sdk-version: " <> projVers)
             createDirectory expected2
             writeFileUTF8 (expected2 </> sdkConfigName) ("version: " <> expected1 <> "\n")
-            (Just (SdkVersion got1), Just (SdkPath got2)) <-
+            (Just got1, Just (SdkPath got2)) <-
                 withEnv [ (sdkVersionEnvVar, Nothing)
                         , (sdkPathEnvVar, Just expected2)
                         ] (getSdk damlPath projectPath)
-            Tasty.assertEqual "sdk version" expected1 (V.toString got1)
+            Tasty.assertEqual "sdk version" expected1 (versionToString got1)
             Tasty.assertEqual "sdk path" expected2 got2
 
     , Tasty.testCase "getSdk: DAML_SDK_VERSION overrides project config version" $ do
@@ -243,11 +243,11 @@ testGetSdk = Tasty.testGroup "DAML.Assistant.Env.getSdk"
             writeFileUTF8 (base </> "project" </> projectConfigName)
                 ("project:\n  sdk-version: " <> projVers)
             createDirectory expected2
-            (Just (SdkVersion got1), Just (SdkPath got2)) <-
+            (Just got1, Just (SdkPath got2)) <-
                 withEnv [ (sdkVersionEnvVar, Just expected1)
                         , (sdkPathEnvVar, Nothing)
                         ] (getSdk damlPath projectPath)
-            Tasty.assertEqual "sdk version" expected1 (V.toString got1)
+            Tasty.assertEqual "sdk version" expected1 (versionToString got1)
             Tasty.assertEqual "sdk path" expected2 got2
 
     , Tasty.testCase "getSdk: Returns Nothings if .daml/sdk is missing." $ do
@@ -298,7 +298,6 @@ testInstall = Tasty.testGroup "DAML.Assistant.Install"
                 options = InstallOptions
                     { iTargetM = Just (RawInstallTarget "source.tar.gz")
                     , iActivate = ActivateInstall True
-                    , iInitial = InitialInstall True
                     , iQuiet = QuietInstall True
                     , iForce = ForceInstall False
                     }
@@ -318,15 +317,19 @@ testInstall = Tasty.testGroup "DAML.Assistant.Install"
                 .| sinkFile "source.tar.gz"
 
             install options damlPath
+    , if isWindows
+        then testInstallWindows
+        else testInstallUnix
+    ]
 
-
-    , Tasty.testCase "reject an absolute symlink in a tarball" $ do
+testInstallUnix :: Tasty.TestTree
+testInstallUnix = Tasty.testGroup "unix-specific tests"
+    [ Tasty.testCase "reject an absolute symlink in a tarball" $ do
         withSystemTempDirectory "test-install" $ \ base -> do
             let damlPath = DamlPath (base </> "daml")
                 options = InstallOptions
                     { iTargetM = Just (RawInstallTarget "source.tar.gz")
                     , iActivate = ActivateInstall False
-                    , iInitial = InitialInstall True
                     , iQuiet = QuietInstall True
                     , iForce = ForceInstall False
                     }
@@ -353,7 +356,6 @@ testInstall = Tasty.testGroup "DAML.Assistant.Install"
                 options = InstallOptions
                     { iTargetM = Just (RawInstallTarget "source.tar.gz")
                     , iActivate = ActivateInstall False
-                    , iInitial = InitialInstall True
                     , iQuiet = QuietInstall True
                     , iForce = ForceInstall False
                     }
@@ -373,6 +375,7 @@ testInstall = Tasty.testGroup "DAML.Assistant.Install"
             assertError "Extracting SDK release tarball."
                 "Invalid SDK release: symbolic link target escapes tarball."
                 (install options damlPath)
-
-
     ]
+
+testInstallWindows :: Tasty.TestTree
+testInstallWindows = Tasty.testGroup "windows-specific tests" []